requirements.txt
# Computer Vision CNC Manufacturing Platform - Core Dependencies

# Deep Learning Frameworks
torch>=1.13.0,<2.0.0
torchvision>=0.14.0,<1.0.0
tensorflow>=2.10.0,<3.0.0
onnx>=1.12.0
onnxruntime>=1.12.0

# Computer Vision
opencv-python>=4.7.0
opencv-contrib-python>=4.7.0
Pillow>=9.3.0
scikit-image>=0.19.0
albumentations>=1.3.0
imageio>=2.22.0

# Machine Learning & AI
scikit-learn>=1.1.0
numpy>=1.21.0,<2.0.0
scipy>=1.9.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Computer Vision Models
detectron2>=0.6
ultralytics>=8.0.0  # YOLOv8
timm>=0.6.0  # PyTorch Image Models

# Industrial Communication Protocols
opcua>=0.98.0
xmltodict>=0.13.0  # For MTConnect XML parsing
pymodbus>=3.0.0
pyserial>=3.5

# Manufacturing Integration
pydantic>=1.10.0
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
requests>=2.28.0
aiohttp>=3.8.0
websockets>=10.4

# Database & Storage
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.8.0
psycopg2-binary>=2.9.0
redis>=4.3.0
pymongo>=4.3.0

# Data Processing
celery[redis]>=5.2.0
apache-airflow>=2.4.0
dask[complete]>=2022.10.0

# Monitoring & Logging
prometheus-client>=0.15.0
structlog>=22.3.0
loguru>=0.6.0
sentry-sdk[fastapi]>=1.10.0

# Security & Compliance
cryptography>=38.0.0
pyjwt>=2.6.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Configuration Management
python-dotenv>=0.21.0
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0

# Testing (moved to requirements-dev.txt)
# Development tools (moved to requirements-dev.txt)

# Performance & Optimization
numba>=0.56.0
cython>=0.29.0
psutil>=5.9.0

# Networking & Communication
pyzmq>=24.0.0
grpcio>=1.50.0
grpcio-tools>=1.50.0

# Time Series & Analytics
influxdb-client>=1.35.0
grafana-api>=1.0.3

# Containerization Support
gunicorn>=20.1.0

# Hardware Interface
# These may need platform-specific installation
# picamera>=1.13  # For Raspberry Pi camera (if needed)
# RPi.GPIO>=0.7.1  # For Raspberry Pi GPIO (if needed)
